requirements.txt
# Aadhaar KYC Face Matching Pipeline — Dependencies
# Install order matters. See CLAUDE.md for step-by-step instructions.
#
# pip install torch==2.3.1+cu121 torchvision==0.18.1+cu121 --extra-index-url https://download.pytorch.org/whl/cu121
# pip install -r requirements.txt

numpy==1.26.4
opencv-python==4.10.0.84
Pillow>=10.4.0
onnxruntime-gpu==1.19.2
basicsr==1.4.2
realesrgan==0.3.0
insightface==0.7.3
PyYAML==6.0.2
requests>=2.32.3
tqdm>=4.66.4
pytest>=8.3.2
